+++ tests/phpunit/tests/linkify-tags.php
@@ -73,7 +73,7 @@ class Linkify_Tags_Test extends WP_UnitTestCase {
 	}
 
 	public function test_widget_version() {
-		$this->assertEquals( '005', c2c_LinkifyTagsWidget::version() );
+		$this->assertEquals( '006', c2c_LinkifyTagsWidget::version() );
 	}
 
 	public function test_widget_hooks_widgets_init() {