+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change: Note compatibility through PHP 8.3+
 * Change: Update copyright date (2025)
 * Unit tests:
+    * Change: Remove vestiges of testing for now-removed `linkify_categories()`
     * Change: Explicitly define return type for overridden methods
 
 ## 2.4 _(2024-08-20)_