+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Hardening: Prevent unsafe markup from being output
 * Update widget base class to v006:
     * Hardening: Prevent unsafe markup from being output
+    * Change: Include version number in class name to ensure use of expected version
     * Change: Move PHPCS-related inline comments with their associated `phpcs:ignore` comments
 * Change: Note compatibility through WP 6.8+
 * Change: Note compatibility through PHP 8.3+